feat(#1076): ICA: worktree-sandbox/shell-tokens single-caller extraction
