+++ .pi/extensions/session-advice/advice-pipeline.ts
@@ -461,7 +461,10 @@ export async function writeAdvice(
 		if (model && modelRegistry) {
 			const auth = await modelRegistry
 				.getApiKeyAndHeaders(model)
-				.catch(() => ({ ok: false, error: "exception" }));
+				.catch((): { ok: boolean; apiKey?: string; error: string } => ({
+					ok: false,
+					error: "exception",
+				}));
 			if (auth.ok && auth.apiKey) {
 				try {
 					llmAdvice = await generateAdvice(analysis, model, modelRegistry);
