+++ .pi/extensions/worktree-sandbox/test/find-unsafe-cd.test.mts
@@ -21,7 +21,8 @@ import { join } from "node:path";
 import { withSandboxEnv, makeCtx } from "./helpers.ts";
 
 // ---- vv BOUNDARY: TDD GATE — do not remove vv ----
-import { findUnsafeCd } from "../index.ts";
+import { findUnsafeCd, SEPARATORS, isCommandStart, findMeaningfulToken } from "../index.ts";
+import type { ParseEntry } from "shell-quote";
 // ---- ^^ BOUNDARY: TDD GATE — do not remove ^^ ----
 
 const SANDBOX_ROOT = "/tmp/sandbox-test-root";
@@ -487,5 +488,73 @@ describe("findUnsafeCd via bash handler (integration)", () => {
 			assert.equal(result.block, true);
 			assert.ok((result.reason ?? "").includes("<previous-dir>"));
 		});
+	});
+});
+
+// ═══════════════════════════════════════════════════════════════════════
+// Ported from shell-tokens.test.mts — edge cases not reachable through
+// guard-level exports (findUnsafeCd, findSuspiciousArg, findUnsafeWriteInBash).
+// These use hand-crafted ParseEntry[] arrays to pin separator/glob/comment
+// behavior that shell-quote.parse() does not naturally produce.
+// ═══════════════════════════════════════════════════════════════════════
+
+describe("isCommandStart (ported from shell-tokens.test.mts)", () => {
+	it("returns true when preceded by ;; separator", () => {
+		const tokens: ParseEntry[] = ["echo", { op: ";;" }, "cat"];
+		assert.equal(isCommandStart(tokens, 2), true);
+	});
+
+	it("returns true when preceded by |& separator", () => {
+		const tokens: ParseEntry[] = ["echo", { op: "|&" }, "cat"];
+		assert.equal(isCommandStart(tokens, 2), true);
+	});
+
+	it("returns true when preceded by & separator", () => {
+		const tokens: ParseEntry[] = ["echo", { op: "&" }, "cat"];
+		assert.equal(isCommandStart(tokens, 2), true);
+	});
+
+	it("returns false when preceded by glob operator", () => {
+		const tokens: ParseEntry[] = ["cmd", { op: "glob", pattern: "*.txt" }];
+		assert.equal(isCommandStart(tokens, 1), false);
+	});
+});
+
+describe("findMeaningfulToken (ported from shell-tokens.test.mts)", () => {
+	it("returns glob when next token is a glob operator", () => {
+		const tokens: ParseEntry[] = ["echo", { op: "glob", pattern: "*.txt" }];
+		const result = findMeaningfulToken(tokens, 1);
+		assert.equal(result.kind, "glob");
+		if (result.kind === "glob") {
+			assert.equal(result.pattern, "*.txt");
+		}
+	});
+
+	it("returns glob at correct index", () => {
+		const tokens: ParseEntry[] = ["cmd", { op: "glob", pattern: "*.ts" }];
+		const result = findMeaningfulToken(tokens, 1);
+		assert.equal(result.kind, "glob");
+		if (result.kind === "glob") {
+			assert.equal(result.index, 1);
+		}
+	});
+
+	it("skips non-separator operators and returns glob when it encounters one", () => {
+		const tokens: ParseEntry[] = [
+			"cmd",
+			{ op: ">" },
+			{ op: "glob", pattern: "*.txt" },
+			{ comment: " note" },
+		];
+		const result = findMeaningfulToken(tokens, 1);
+		assert.equal(result.kind, "glob");
+	});
+});
+
+describe("SEPARATORS (ported from shell-tokens.test.mts)", () => {
+	it("does not contain non-separator operators", () => {
+		assert.equal(SEPARATORS.has(">"), false);
+		assert.equal(SEPARATORS.has(">>"), false);
+		assert.equal(SEPARATORS.has("glob"), false);
 	});
 });
