+++ docs/extensions/worktree-sandbox.md
@@ -24,8 +24,7 @@ nav_order: 7
 Deterministic tool-call interception with shell-aware path analysis:
 
 ```
-├── index.ts        # Entry: tool_call handler, path rewrite, shell-aware escape detection
-├── shell-tokens.ts # isCommandStart, findMeaningfulToken, SEPARATORS set
+├── index.ts        # Entry: tool_call handler, path rewrite, shell-aware escape detection + inlined token helpers
 └── test/           # Unit tests for all enforcement paths
 ```
